refactor: inline pairing and genuine-check fragments, replace onNFCAvailableRef with an option (#242)
usePairing and useGenuineCheck each exported five fields to exactly one
consumer and could not complete their own protocols without the coordinator
injecting startNFC. Inline both into useKeycardOperation: the onRestart
contracts disappear, the phase-priority expression reads locally owned
state, and the ADR-0005 pairing-password flow is byte-for-byte unchanged.

useNFCSession no longer exposes a mutable onNFCAvailableRef that a caller
two layers up had to assign during render. It takes an onNFCAvailable
option at construction; useNFCOperation forwards it, and the coordinator
passes its own retry(), so the NFC-re-enabled reaction lives in one place.
Also deletes the unused clearPinError from the coordinator interface and
fixes the long-standing type errors in useFactoryReset and useInitCard,
which returned objects missing openNFCSettings.
